+++ client/src/components/pages/RiderPage.jsx
@@ -1,0 +1,20 @@
+import { useEffect } from 'react';
+import { Navigate } from 'react-router';
+
+import RiderContent from '../riders/RiderContent';
+
+const Rider = () => {
+  const rider = JSON.parse(localStorage.getItem('rider'));
+
+  useEffect(() => {
+    document.title = 'Rider';
+  }, []);
+
+  if (!rider) {
+    return  <Navigate to={'/'} replace />;
+  }
+
+  return <RiderContent riderId={rider._id} riderName={rider.name} />;
+};
+
+export default Rider;
